Convert the description of the report descriptor into a C-compatible one
The idea is to make use of https://gitlab.freedesktop.org/libevdev/udev-hid-bpf/-/blob/main/src/bpf/hid_report_helpers.h
The description doesn't lose much of its value from a human perspective
but can now be parsed by a C compiler if we include that file from above.

It's probably not perfect, but it seems to give correct results for
the device I tested it with.